requirements.txt
streamlit>=1.42,<2.0
google-genai>=1.0
Pillow>=10.0
requests>=2.31
python-dotenv>=1.0
